Almost a prime number and a brilliant number
In this test, I am using a web application that includes html, css and js.The app will show interesting numbers. Almost a prime number and a brilliant number *
The user will enter in the txtInput text box any positive integer. And press one of the buttons.If the input is invalid, you will receive an error message as *
 a pop-up. Fill out the pop-up window as a new div that appears on the screen with the OK button and after the user clicks on it, it will disappear.          *
If the input is valid, all the near-primary and non-bright numbers in the range will be calculated between 1 and the number entered (we will denote it as n). *
The results will be displayed in Area D( Table shape)